+++ src/cpshell.cpp
@@ -104,7 +104,7 @@ int cpshell_main(int argc, char **argv)
         while (a->name[0] != 0) {
             strcpy(cmds, (a++)->name);
             // check if terminal->bufferCX is at the end of the line + 2 for ', '
-            if ((terminal->bufferCX + strlen(a->name) + 2) >= terminal->xmax) {
+            if ((terminal->bufferCX + (int16_t)(strlen(a->name) + 2) >= terminal->xmax) {
                 terminal->WriteBuffer('\n', false);
                 terminal->ClearBuffer();
             }
